refactor: move series scales to types/xy_chart/utils
