+++ src/state/test/interactions.test.ts
@@ -1,5 +1,5 @@
 import { BarGeometry } from '../../types/xy_chart/rendering/rendering';
-import { computeXScale, computeYScales } from '../../lib/series/scales';
+import { computeXScale, computeYScales } from '../../types/xy_chart/utils/scales';
 import { DataSeriesColorsValues } from '../../lib/series/series';
 import { BarSeriesSpec, BasicSeriesSpec, RectAnnotationSpec } from '../../lib/series/specs';
 import { getAnnotationId, getGroupId, getSpecId } from '../../utils/ids';
