+++ src/types/xy_chart/rendering/rendering.bars.test.ts
@@ -3,7 +3,7 @@ import { identity } from '../../../utils/commons';
 import { getGroupId, getSpecId, SpecId } from '../../../utils/ids';
 import { ScaleType } from '../../../utils/scales/scales';
 import { renderBars } from './rendering';
-import { computeXScale, computeYScales } from '../../../lib/series/scales';
+import { computeXScale, computeYScales } from '../utils/scales';
 import { BarSeriesSpec } from '../../../lib/series/specs';
 import { LIGHT_THEME } from '../../../utils/themes/light_theme';
