+++ src/types/xy_chart/rendering/rendering.lines.test.ts
@@ -5,7 +5,7 @@ import { getGroupId, getSpecId, SpecId } from '../../../utils/ids';
 import { ScaleType } from '../../../utils/scales/scales';
 import { CurveType } from '../../../utils/curves';
 import { IndexedGeometry, LineGeometry, PointGeometry, renderLine } from './rendering';
-import { computeXScale, computeYScales } from '../../../lib/series/scales';
+import { computeXScale, computeYScales } from '../utils/scales';
 import { LineSeriesSpec } from '../../../lib/series/specs';
 import { LIGHT_THEME } from '../../../utils/themes/light_theme';
