+++ src/types/xy_chart/utils/axis_utils.ts
@@ -1,6 +1,6 @@
 import { XDomain } from '../../../lib/series/domains/x_domain';
 import { YDomain } from '../../../lib/series/domains/y_domain';
-import { computeXScale, computeYScales } from '../../../lib/series/scales';
+import { computeXScale, computeYScales } from './scales';
 import {
   AxisSpec,
   CompleteBoundedDomain,
