+++ src/types/xy_chart/utils/scales.test.ts
@@ -1,8 +1,8 @@
-import { getGroupId } from '../../utils/ids';
-import { ScaleType } from '../../utils/scales/scales';
-import { XDomain } from './domains/x_domain';
+import { getGroupId } from '../../../utils/ids';
+import { ScaleType } from '../../../utils/scales/scales';
+import { XDomain } from '../../../lib/series/domains/x_domain';
 import { computeXScale, countBarsInCluster } from './scales';
-import { FormattedDataSeries } from './series';
+import { FormattedDataSeries } from '../../../lib/series/series';
 
 describe('Series scales', () => {
   const xDomainLinear: XDomain = {
