+++ src/types/xy_chart/utils/scales.ts
@@ -1,10 +1,10 @@
-import { GroupId } from '../../utils/ids';
-import { ScaleBand } from '../../utils/scales/scale_band';
-import { ScaleContinuous } from '../../utils/scales/scale_continuous';
-import { Scale, ScaleType } from '../../utils/scales/scales';
-import { XDomain } from './domains/x_domain';
-import { YDomain } from './domains/y_domain';
-import { FormattedDataSeries } from './series';
+import { GroupId } from '../../../utils/ids';
+import { ScaleBand } from '../../../utils/scales/scale_band';
+import { ScaleContinuous } from '../../../utils/scales/scale_continuous';
+import { Scale, ScaleType } from '../../../utils/scales/scales';
+import { XDomain } from '../../../lib/series/domains/x_domain';
+import { YDomain } from '../../../lib/series/domains/y_domain';
+import { FormattedDataSeries } from '../../../lib/series/series';
 
 /**
  * Count the max number of bars in cluster value.
